Use cbor for Inbound/Outbound codec
